pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.1",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-openai>=0.3.17",
    "markdown>=3.8",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
]
